Inventory now Dictionary
Changed the list in inventory to a dictionary. The dictionary will hold the name of the item and the number of times one was picked up. I added a name value to the cup object and renamed the object to trash. Added the set name function to the trash object.